+++ beacon-chain/blockchain/service_test.go
@@ -446,47 +446,3 @@ func TestIsBlockReadyForProcessing(t *testing.T) {
 	}
 
 }
-
-func TestUpdateBlockVoteCache(t *testing.T) {
-	db := internal.SetupDB(t)
-	defer internal.TeardownDB(t, db)
-	chainService := setupBeaconChain(t, true, db)
-
-	beaconState, err := types.NewGenesisBeaconState(nil)
-	if err != nil {
-		t.Fatalf("failed to initialize genesis state: %v", err)
-	}
-	block := types.NewBlock(&pb.BeaconBlock{
-		Slot:             1,
-		ParentRootHash32: []byte{},
-		Attestations: []*pb.AggregatedAttestation{
-			{
-				Slot:             0,
-				Shard:            1,
-				AttesterBitfield: []byte{'F', 'F'},
-			},
-		},
-	})
-
-	err = chainService.calculateNewBlockVotes(block, beaconState)
-	if err != nil {
-		t.Errorf("failed to update the block vote cache: %v", err)
-	}
-}
-
-func TestUpdateBlockVoteCacheNoAttestations(t *testing.T) {
-	db := internal.SetupDB(t)
-	defer internal.TeardownDB(t, db)
-	chainService := setupBeaconChain(t, true, db)
-
-	beaconState, err := types.NewGenesisBeaconState(nil)
-	if err != nil {
-		t.Fatalf("failed to initialize genesis state: %v", err)
-	}
-	block := types.NewBlock(nil)
-
-	err = chainService.calculateNewBlockVotes(block, beaconState)
-	if err != nil {
-		t.Errorf("failed to update the block vote cache: %v", err)
-	}
-}
