+++ beacon-chain/core/incentives/BUILD.bazel
@@ -6,10 +6,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/incentives",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//beacon-chain/core/validators:go_default_library",
-        "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bitutil:go_default_library",
         "//shared/mathutil:go_default_library",
         "//shared/params:go_default_library",
     ],
