+++ beacon-chain/core/incentives/incentives_test.go
@@ -18,93 +18,6 @@ func newValidatorRegistry() []*pb.ValidatorRecord {
 	return validators
 }
 
-func TestComputeValidatorRewardsAndPenalties(t *testing.T) {
-	validators := newValidatorRegistry()
-	defaultBalance := uint64(32 * 1e9)
-
-	participatedDeposit := 4 * defaultBalance
-	totalDeposit := 10 * defaultBalance
-	rewQuotient := RewardQuotient(totalDeposit)
-	penaltyQuotient := QuadraticPenaltyQuotient()
-	timeSinceFinality := uint64(5)
-
-	data := &pb.BeaconState{
-		ValidatorRegistry:               validators,
-		ValidatorRegistryLastChangeSlot: 1,
-		JustifiedSlot:                   4,
-		FinalizedSlot:                   3,
-	}
-
-	activeValidatorIndices := make([]uint32, 0, len(validators))
-	for i, v := range validators {
-		if v.Status == pb.ValidatorRecord_ACTIVE {
-			activeValidatorIndices = append(activeValidatorIndices, uint32(i))
-		}
-	}
-
-	rewardedValidatorRegistry := CalculateRewards(
-		[]uint32{2, 3, 6, 9},
-		activeValidatorIndices,
-		data.ValidatorRegistry,
-		totalDeposit,
-		participatedDeposit,
-		timeSinceFinality,
-	)
-
-	expectedBalance := defaultBalance - defaultBalance/uint64(rewQuotient)
-
-	if rewardedValidatorRegistry[0].Balance != expectedBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[0].Balance, expectedBalance)
-	}
-
-	expectedBalance = uint64(int64(defaultBalance) + int64(defaultBalance/rewQuotient)*(2*int64(participatedDeposit)-int64(totalDeposit))/int64(totalDeposit))
-
-	if rewardedValidatorRegistry[6].Balance != expectedBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[6].Balance, expectedBalance)
-	}
-
-	if rewardedValidatorRegistry[9].Balance != expectedBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[9].Balance, expectedBalance)
-	}
-
-	validators = newValidatorRegistry()
-	timeSinceFinality = 200
-
-	activeValidatorIndices = make([]uint32, 0, len(validators))
-	for i, v := range validators {
-		if v.Status == pb.ValidatorRecord_ACTIVE {
-			activeValidatorIndices = append(activeValidatorIndices, uint32(i))
-		}
-	}
-
-	rewardedValidatorRegistry = CalculateRewards(
-		[]uint32{1, 2, 7, 8},
-		activeValidatorIndices,
-		validators,
-		totalDeposit,
-		participatedDeposit,
-		timeSinceFinality)
-
-	if rewardedValidatorRegistry[1].Balance != defaultBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[1].Balance, defaultBalance)
-	}
-
-	if rewardedValidatorRegistry[7].Balance != defaultBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[7].Balance, defaultBalance)
-	}
-
-	expectedBalance = defaultBalance - (defaultBalance/rewQuotient + defaultBalance*timeSinceFinality/penaltyQuotient)
-
-	if rewardedValidatorRegistry[0].Balance != expectedBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[0].Balance, expectedBalance)
-	}
-
-	if rewardedValidatorRegistry[9].Balance != expectedBalance {
-		t.Fatalf("validator balance not updated correctly: %d, %d", rewardedValidatorRegistry[9].Balance, expectedBalance)
-	}
-
-}
-
 func TestRewardQuotient(t *testing.T) {
 	defaultBalance := uint64(2 * 1e9)
 	totalDeposit := defaultBalance
