+++ beacon-chain/core/state/BUILD.bazel
@@ -2,46 +2,23 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "processing.go",
-        "state_transition.go",
-        "validity_conditions.go",
-    ],
+    srcs = ["state_transition.go"],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/state",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//beacon-chain/core/incentives:go_default_library",
         "//beacon-chain/core/randao:go_default_library",
         "//beacon-chain/core/types:go_default_library",
-        "//beacon-chain/core/validators:go_default_library",
-        "//beacon-chain/utils:go_default_library",
-        "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bitutil:go_default_library",
-        "//shared/hashutil:go_default_library",
         "//shared/params:go_default_library",
-        "@com_github_ethereum_go_ethereum//common:go_default_library",
-        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "processing_test.go",
-        "state_transition_test.go",
-        "validity_conditions_test.go",
-    ],
+    srcs = ["state_transition_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/types:go_default_library",
-        "//beacon-chain/core/validators:go_default_library",
-        "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bytes:go_default_library",
-        "//shared/hashutil:go_default_library",
         "//shared/params:go_default_library",
-        "@com_github_ethereum_go_ethereum//common:go_default_library",
-        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
